Refactor container permission contracts and docs
