+++ server/containerPermissions/index.ts
@@ -1,0 +1,6 @@
+export {
+  applyContainerPermissionChangeSet,
+  applyContainerPermissions,
+  fetchContainerPermissionEntries,
+  listContainerPermissions,
+} from "./containerPermissionsHandlers";
