+++ src/components/permissions/services/directoryPrincipalSearch/directoryPrincipalSearchTypes.ts
@@ -1,4 +1,4 @@
-import { PermissionTabValue } from "../../models/permissionModels";
+import type { PermissionTabValue } from "../../models/permissionModels";
 
 /**
  * 目录搜索支持的 principal 类型。
