+++ tsconfig.json
@@ -24,5 +24,5 @@
     "isolatedModules": true,
     "jsx": "react-jsx"
   },
-  "include": ["src"]
+  "include": ["src", "common/contracts"]
 }
